% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_to.R
\name{note_to_title}
\alias{note_to_title}
\alias{note_to_source_note}
\alias{title_to_notes}
\alias{source_to_notes}
\alias{title_source_to_notes}
\title{Move notes to and from the title/source note of a tibble}
\usage{
note_to_title(x, note, .overwrite = FALSE)

note_to_source_note(x, note, .overwrite = FALSE)

title_to_notes(x, .add_before = 0)

source_to_notes(x, .add_before = Inf)

title_source_to_notes(x, .add_before = 0)
}
\arguments{
\item{x}{A \code{shrthnd_tbl()} object}

\item{note}{The number of the note to move}

\item{.overwrite}{Whether to overwrite existing}

\item{.add_before}{Where to (re)insert the note}
}
\value{
A \code{shrthnd_tbl()}
}
\description{
A \code{shrthnd_tbl()} has three sets of \link{annotations}, the
\verb{note_to_*()} functions allow you to move a general note to either the
title or source note of a tibble. The \verb{*_to_notes()}functions do the
opposite and (re)insert either the title and/or source note back into the
general notes.
}
\details{
For \code{title_to_notes()} and \code{title_source_to_notes()} the default is to
(re)insert the note at the start of the set of notes, for
\code{source_to_notes()} the default is to (re)insert the note at the end of
the set of notes.
}
\examples{
x <- c("12", "34.567", "[c]", "NA", "56.78[e]", "78.9", "90.123[e]")
sh_x <- shrthnd_num(x, c("[c]", "[e]"))
tbl <- tibble::tibble(x = x, sh_x = sh_x)

sh_tbl <- shrthnd_tbl(tbl) |>
  set_notes(c("Note 1", "Note 2", "Note 3")) |>
  note_to_title(1) |>
  note_to_source_note(2)

sh_tbl

shrthnd_notes(sh_tbl)

sh_tbl <- sh_tbl |>
  title_to_notes()

shrthnd_notes(sh_tbl)

}
\seealso{

\code{\link{annotations}()},
\code{\link{shrthnd_tbl}()},
\code{\link{zap_shrthnd}()}
}
\concept{tbl}
