% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{annotations}
\alias{annotations}
\alias{shrthnd_title}
\alias{shrthnd_title<-}
\alias{set_title}
\alias{shrthnd_source_note}
\alias{shrthnd_source_note<-}
\alias{set_source_note}
\alias{shrthnd_notes}
\alias{shrthnd_notes<-}
\alias{set_notes}
\alias{add_notes}
\alias{add_notes<-}
\alias{set_tbl_antn}
\title{Get and set annotations for a tibble}
\usage{
annotations(x)

shrthnd_title(x)

shrthnd_title(x) <- value

set_title(x, value, .overwrite = FALSE)

shrthnd_source_note(x)

shrthnd_source_note(x) <- value

set_source_note(x, value, .overwrite = FALSE)

shrthnd_notes(x)

shrthnd_notes(x) <- value

set_notes(x, value, .overwrite = FALSE)

add_notes(x, value, .add_before = Inf)

add_notes(x) <- value

set_tbl_antn(
  x,
  what = c("title", "source_note", "notes"),
  value,
  .overwrite = FALSE,
  .add = FALSE,
  .add_before = Inf
)
}
\arguments{
\item{x}{A \code{shrthnd_tbl()} object}

\item{value}{The value to set}

\item{.overwrite}{Whether an existing value should be overwritten}

\item{.add_before}{When adding notes, where to add the note (defaults to
the end of the current set of notes)}

\item{what}{Which note to set, one of \code{title}, \code{source_note} or \code{notes}}

\item{.add}{When \code{what = "notes"}, whether to append to the existing
set of notes}
}
\value{
For \code{shrthnd_title()}, \code{shrthnd_source_note()} and \code{shrthnd_notes()}
a character vector of the note(s). For the setting functions returns
invisibly either \code{x} if the attribute was set or \code{NULL} if not.
}
\description{
A \code{shrthnd_tbl()} has three sets of annotations that can be defined: a
\code{title}, a \code{source_note} and a set of general \code{notes}. This family of
functions allows you to view and modify these notes.
}
\details{
Use \code{annotations()} to see the all the annotations associated with a
\code{shrthnd_tbl()} object.

Use \code{shrthnd_title()}, \code{shrthnd_source_note()} and \code{shrthnd_notes()} get the
relevant annotations(s) of a \code{shrthnd_tbl()} object. Passing a value to these
functions (e.g. \code{shrthnd_title(x) <- "My title"}) will set the value of these
annotation, overwriting the existing value(s).

\code{set_title()}, \code{set_source_note()}, and \code{set_notes()} also allow you to set
the value of these annotations. By default they will not permit overwriting
of existing values, setting \code{.overwrite = TRUE} permits this.

\code{add_notes()} allows you to append notes to the existing set of general
notes.

\code{set_tbl_antn()} is a low level helper function that powers the assignment
operations.
}
\examples{
x <- c("12", "34.567", "[c]", "NA", "56.78[e]", "78.9", "90.123[e]")
sh_x <- shrthnd_num(x, c("[c]", "[e]"))
tbl <- tibble::tibble(x = x, sh_x = sh_x)

sh_tbl <- shrthnd_tbl(tbl) |>
  set_title("My Example Table") |>
  set_source_note("Shrthnd documentation (2023)") |>
  set_notes(c("Note 1", "Note 2"))

sh_tbl

annotations(sh_tbl)

shrthnd_title(sh_tbl)
shrthnd_source_note(sh_tbl)
shrthnd_notes(sh_tbl)

add_notes(sh_tbl) <- "Note 3"
shrthnd_notes(sh_tbl)

}
\seealso{

\code{\link{note_to_title}()},
\code{\link{shrthnd_tbl}()},
\code{\link{zap_shrthnd}()}
}
\concept{tbl}
