% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_to.R
\name{note_to_title}
\alias{note_to_title}
\alias{note_to_source_note}
\title{Convert general notes to tibble title or source note}
\usage{
note_to_title(x, note, .overwrite = FALSE)

note_to_source_note(x, note, .overwrite = FALSE)
}
\arguments{
\item{x}{A \code{shrthnd_tbl()} object}

\item{note}{The number of the note to move}

\item{.overwrite}{Whether to overwrite existing}
}
\value{
A \code{shrthnd_tbl()}
}
\description{
The \code{note_to_title()} and \code{note_to_source_note()} allow you to move a note
from inside the general \code{shrthnd_notes()} of a \code{shrthnd_tbl()} and move
them to the \code{shrthnd_title()} or \code{shrthnd_source_note()} of the tibble.
}
\examples{
x <- c("12", "34.567", "[c]", "NA", "56.78[e]", "78.9", "90.123[e]")
sh_x <- shrthnd_num(x, c("[c]", "[e]"))
tbl <- tibble::tibble(x = x, sh_x = sh_x)

sh_tbl <- shrthnd_tbl(tbl) |>
  set_notes(c("Note 1", "Note 2", "Note 3")) |>
  note_to_title(1) |>
  note_to_source_note(2)

sh_tbl

shrthnd_notes(sh_tbl)
}
