% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{shrthnd_coercion}
\alias{shrthnd_coercion}
\title{Coercion of shrthnd_num vectors}
\description{
Through the \code{{vectrs}} package a \code{shrthnd_num()} is generally coerced to
behave as if it was a regular a \code{numeric()} vector. Where \code{{vectrs}}
doesn't automatically support coercion custom methods are provided to
enable a \code{shrthnd_num()} to be considered as a numeric vector.
}
\details{
Of particular note is that \code{is.na()} will return \code{TRUE} if the numeric
component is missing, i.e. if there is no numeric element but a tag
marker then \code{is.na()} will return \code{TRUE}. Use \code{is_na_tag()} to identify
where there is no tag marker, or \code{is_na_both} to identify where both the
numeric and tag components are missing.

See \code{\link{shrthnd_maths}} for details on how \code{shrthnd_num()} works with
arithmetic and mathematical operations.
}
\seealso{

\code{\link{as_shrthnd}()},
\code{\link{is_shrthnd_num}()},
\code{\link{make_shrthnd_num}()},
\code{\link{shrthnd_maths}},
\code{\link{shrthnd_num}()},
\code{\link{shrthnd_tags}()},
\code{\link{tag_match}()}
}
\concept{num}
