% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrthand_list.R
\name{shrthnd_list}
\alias{shrthnd_list}
\title{Collect the shorthand in a vector}
\usage{
shrthnd_list(x, shrthnd_values, na_values = "NA")
}
\arguments{
\item{x}{A character vector containing shorthand}

\item{na_values}{A character value of NA values}

\item{shorthand}{A character vector of shorthand values}
}
\value{
A list of shorthand positions in a vector
}
\description{
Collect the shorthand in a vector
}
\examples{
x <- c("12.345", "34.567", "[c]", "NA", "56.78[e]", "78.9", "90.123[e]")
shrthnd_list(x, c("[c]", "[e]"))
}
