% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths.R
\name{shrthnd_maths}
\alias{shrthnd_maths}
\title{Arithmetic and mathematical operations}
\description{
Arithmetic and most mathematical operations are supported on the numeric
component of \code{shrthnd_num()} vectors via the \code{{vctrs}} package without
having to wrap the vector in \code{as.numeric()}.
}
\details{
You can use all the standard arithmetic infix operators (\code{+}, \code{-}, \code{/}, \code{*},
\code{^}, \code{\%\%}, \code{\%/\%}, \code{!}). See \code{vctrs::vec_arith()} for further details.

Through \code{vctrs::vec_math()} the following generic mathematical operations
are supported:
\itemize{
\item from the \link{Summary} group generic:
\code{prod()}, \code{sum()}, \code{any()}, \code{all()}.
\item from the \link{Math} group generic:
\code{abs()}, \code{sign()}, \code{sqrt()}, \code{ceiling()}, \code{floor()}, \code{trunc()}, \code{cummax()},
\code{cummin()}, \code{cumprod()}, \code{cumsum()}, \code{log()}, \code{log10()}, \code{log2()},
\code{log1p()}, \code{acos()}, \code{acosh()}, \code{asin()}, \code{asinh()}, \code{atan()}, \code{atanh()},
\code{exp()}, \code{expm1()}, \code{cos()}, \code{cosh()}, \code{cospi()}, \code{sin()}, \code{sinh()},
\code{sinpi()}, \code{tan()}, \code{tanh()}, \code{tanpi()}, \code{gamma()}, \code{lgamma()},
\code{digamma()}, \code{trigamma()}.
\item \code{vctrs::vec_math()} also enables support for \code{mean()}, \code{is.nan()},
\code{is.finite()} and \code{is.infinite()}.
\item In addition to these, the \code{{shrthnd}} package also provides methods for
\code{median()} and \code{quantile()}, and will work with \code{sd()} due to the ability
of a \code{shrthnd_num()} to be easily coerced to a numeric vector.
}

For other operations you will need to wrap the \code{shrthnd_num} vector in
\code{as.numeric()}.

For all operations remember that you will likely need to set \code{na.rm = TRUE}
or whatever other method a function has for ignoring missing values.
}
\examples{
x <- c("12", "34.567", "[c]", "NA", "56.78[e]", "78.9", "90.123[e]")
sh_x <- shrthnd_num(x, c("[c]", "[e]"))

sh_x * 2

2 + sh_x

mean(sh_x, na.rm = TRUE)

}
\seealso{

\code{\link{as_shrthnd}()},
\code{\link{is_shrthnd_num}()},
\code{\link{make_shrthnd_num}()},
\code{\link{shrthnd_coercion}},
\code{\link{shrthnd_num}()},
\code{\link{shrthnd_tags}()},
\code{\link{tag_match}()}
}
\concept{num}
