% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrthnd_tags.R
\name{shrthnd_tags}
\alias{shrthnd_tags}
\alias{shrthnd_unique_tags}
\title{Get the tags attached to a shrthnd vector}
\usage{
shrthnd_tags(x)

shrthnd_unique_tags(x)
}
\arguments{
\item{x}{A \code{shrthnd_num()} vector}
}
\value{
A character vector
}
\description{
\code{shrthnd_tags()} provides a character vector the same length as \code{x} with
the shorthand tags, or \code{NA} if that value has no tag. \code{shrthnd_unique_tags()}
is a convenience wrapper for \code{unique(shrthnd_tags(x))}, but can also be
called on a \code{shrthnd_list()} object.
}
\examples{
x <- c("12", "34.567", "[c]", "NA", "56.78[e]", "78.9", "90.123[e]")
sh_x <- shrthnd_num(x, c("[c]", "[e]"))
shrthnd_tags(sh_x)
shrthnd_unique_tags(sh_x)
}
