% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{shrthnd_tbl}
\alias{shrthnd_tbl}
\title{Add annotations to tibbles}
\usage{
shrthnd_tbl(tbl, title = NULL, notes = NULL, source_note = NULL)
}
\arguments{
\item{tbl}{A \code{tibble::tibble()} or object that can be coerced to a tibble.}

\item{title}{A character vector for the title of \code{tbl}}

\item{notes}{A character vector of general notes relating to \code{tbl}}

\item{source_note}{A character vector for a source note relating to \code{tbl}}
}
\value{
A tibble with shrthnd annotations
}
\description{
\code{shrthnd_tbl()} provides a way to attach annotations to a table.
Specifically, it supports three types of annotation: a \code{title}, a
\verb{source note} and general \code{notes}. The \code{title} and \code{source_note} are each
character vectors of length 1, while \code{notes} can be a character vector of
any length.
}
\examples{
x <- c("12", "34.567", "[c]", "NA", "56.78[e]", "78.9", "90.123[e]")
sh_x <- shrthnd_num(x, c("[c]", "[e]"))
tbl <- tibble::tibble(x = x, sh_x = sh_x)
shrthnd_tbl(
  tbl,
  title = "Example table",
  notes = c("Note 1", "Note 2"),
  source_note = "Shrthnd documentation, 2023"
)
}
\seealso{
\code{\link[=is_shrthnd_tbl]{is_shrthnd_tbl()}}


\code{\link{annotations}()},
\code{\link{note_to_title}()},
\code{\link{zap_shrthnd}()}
}
\concept{tbl}
