% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrthand_values.R
\name{get_shrthnd_list}
\alias{get_shrthnd_list}
\alias{shrthnd_tags}
\title{Get the shorthand values attached to a vector}
\usage{
get_shrthnd_list(x)

shrthnd_tags(x)
}
\arguments{
\item{x}{A \code{\link[=shrthnd_num]{shrthnd_num()}} vector}
}
\value{
The shrthnd_list associated with x or a character vector of
shorthand tags the same length as the input
}
\description{
Get the shorthand values attached to a vector
}
\examples{
x <- c("12.345", "34.567", "[c]", "NA", "56.78[e]", "78.9", "90.123[e]")
sh_x <- shrthnd_num(x, c("[c]", "[e]"))
get_shrthnd_list(sh_x)
shrthnd_tags(sh_x)
}
