% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{where_shrthnd_cols}
\alias{where_shrthnd_cols}
\alias{which_shrthnd_cols}
\alias{any_shrthnd_cols}
\title{Identify which columns use shrthnd}
\usage{
where_shrthnd_cols(x)

which_shrthnd_cols(x, .names = TRUE)

any_shrthnd_cols(x)
}
\arguments{
\item{x}{A data.frame (or list)}

\item{.names}{A logical vector indicating whether to return column names
(the default) or an integer vector of column positions}
}
\value{
For \code{where_shrthnd_cols()} a logical vector of the same length as
the number columns in \code{x}. For \code{which_shrthnd_cols()} a character vector
of names (the default) or an integer vector of index positions for the
columns (or elements) that are shrthnd_num vectors. For
\code{any_shrthnd_cols()} either \code{TRUE} if there are any \code{shrthnd_num()}
vectors in the object or \code{FALSE} if not.
}
\description{
\code{where_shrthnd_cols()} applies \code{is_shrthnd_num()} across columns of a
data.frame (or elements in a list). \code{which_shrthnd_cols()} identifies the
columns in a data.frame or (elements of a list) by name or index position.
\code{any_shrthnd_cols()} tests whether a data.frame (or list) has any columns
that are \code{shrthnd_num()} vectors.
}
\examples{
x <- c("12", "34.567", "[c]", "NA", "56.78[e]", "78.9", "90.123[e]")
sh_x <- shrthnd_num(x, c("[c]", "[e]"))
tbl <- tibble::tibble(x = x, sh_x = sh_x)

where_shrthnd_cols(tbl)

which_shrthnd_cols(tbl)

which_shrthnd_cols(tbl, .names = FALSE)

any_shrthnd_cols(tbl)
}
