% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zap.R
\name{zap_shrthnd}
\alias{zap_shrthnd}
\alias{zap_title}
\alias{zap_source_note}
\alias{zap_notes}
\alias{zap_tbl}
\title{Remove annotations from tibbles}
\usage{
zap_title(x)

zap_source_note(x)

zap_notes(x)

zap_tbl(x)

zap_shrthnd(x, what = c("title", "source_note", "notes"), zap_class = FALSE)
}
\arguments{
\item{x}{A \code{shrthnd_tbl()}}

\item{what}{One or more of \code{title}, \code{source_note} or \code{notes} indicating
which set of notes to remove}

\item{zap_class}{Whether to remove the "shrthnd_tbl" class}
}
\value{
Returns \code{x} with relevant attributes removed
}
\description{
The \verb{zap_*()} functions remove annotations from a \code{shrthnd_tbl()} object.
\code{zap_title()}, \code{zap_source_note()} and \code{zap_notes()} remove the title,
source note and general notes respectively. \code{zap_tbl()} removes all three
types of annotations and also strips the \code{shrthnd_tbl} class from the object.
\code{zap_shrthnd()} is a low-level helper function that power the attribute
removal.
}
\details{
To remove shrthnd from a \code{shrthnd_num()} vector use
\code{as.numeric()}, \code{as.character()} or \code{as_shrthnd()} to coerce the vector to
another type.
}
\examples{
x <- c("12", "34.567", "[c]", "NA", "56.78[e]", "78.9", "90.123[e]")
sh_x <- shrthnd_num(x, c("[c]", "[e]"))
tbl <- tibble::tibble(x = x, sh_x = sh_x)
sh_tbl <- shrthnd_tbl(
  tbl,
  title = "Example table",
  notes = c("Note 1", "Note 2"),
  source_note = "Shrthnd documentation, 2023"
)

sh_tbl

zap_title(sh_tbl)

zap_source_note(sh_tbl)

zap_notes(sh_tbl)

zap_tbl(sh_tbl)

}
\seealso{

\code{\link{annotations}()},
\code{\link{note_to_title}()},
\code{\link{shrthnd_tbl}()}
}
\concept{tbl}
