% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_shrthnd.R
\name{as_shrthnd}
\alias{as_shrthnd}
\title{Coerce a shrthnd_num to a character vector with shorthand}
\usage{
as_shrthnd(x, digits = TRUE, .subtle = FALSE)
}
\arguments{
\item{x}{A \code{\link[=shrthnd_num]{shrthnd_num()}} vector}

\item{digits}{Whether to apply digit formatting}

\item{.subtle}{Used for formatting with the \code{{pillar}} package}
}
\value{
A character vector
}
\description{
Coerce a shrthnd_num to a character vector with shorthand
}
\examples{
x <- c("12.345", "34.567", "[c]", "NA", "56.78[e]", "78.9", "90.123[e]")
sh_x <- shrthnd_num(x, c("[c]", "[e]"), digits = 1)
as_shrthnd(sh_x)
as_shrthnd(sh_x, digits = FALSE)
}
