% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_shrthnd.R
\name{is_shrthnd_num}
\alias{is_shrthnd_num}
\alias{is_shrthnd}
\alias{is_shrthnd_integer}
\alias{is_shrthnd_double}
\alias{is_shrthnd_list}
\title{shrthnd predicates}
\usage{
is_shrthnd_num(x)

is_shrthnd_integer(x)

is_shrthnd_double(x)

is_shrthnd_list(x)
}
\arguments{
\item{x}{An object to be tested}
}
\value{
A logical vector
}
\description{
The \verb{is_shrthnd_*} family of functions test whether a vector is either a
\code{shrthnd_num()}, or a \code{shrthnd_list()}. \code{is_shrthnd_integer()} and
\code{is_shrthnd_double()} test whether an object is a \code{shrthnd_num()} vector
and whether the underlying data type is an \code{integer()} or a \code{double()}
}
\examples{
x <- c("12", "34.567", "[c]", "NA", "56.78[e]", "78.9", "90.123[e]")
sh_x <- shrthnd_num(x, c("[c]", "[e]"))
is_shrthnd_num(sh_x)
is_shrthnd_double(sh_x)

y <- c("12", "34", "[c]", "NA", "56[e]", "78", "90[e]")
sh_y <- shrthnd_num(y, c("[c]", "[e]"))
is_shrthnd_num(sh_y)
is_shrthnd_integer(sh_y)

sh_l <- shrthnd_list(sh_x)
is_shrthnd_list(sh_l)
}
