% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrthnd_num.R
\name{make_shrthnd_num}
\alias{make_shrthnd_num}
\title{Make a shrthnd_num vector}
\usage{
make_shrthnd_num(x = numeric(), tags = character(), digits = 2L)
}
\arguments{
\item{x}{A numeric vector}

\item{tags}{A character vector}

\item{digits}{The number of digits to format the numeric vector with}
}
\value{
A \code{shrthnd_num} vector
}
\description{
\code{make_shrthnd_num()} allows you to construct a \code{shrthnd_num} vector from
a numeric vector of data values and a character vector of shorthand markers.
}
\examples{
make_shrthnd_num(c(1:3, NA, 4:5, NA), c("", "", "", "[c]", "", "[e]", NA))
}
