% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{shrthnd_list}
\alias{shrthnd_list}
\title{List the shorthand in a vector}
\usage{
shrthnd_list(
  x,
  shorthand = NULL,
  na_values = c("", "NA"),
  dec = ".",
  bigmark = ","
)
}
\arguments{
\item{x}{A character vector containing shorthand, or a \code{shrthnd_num()} vector}

\item{shorthand}{A character vector of shorthand values to validate tags against}

\item{na_values}{A character value of NA values to ignore}

\item{dec}{The decimal separator for numbers}

\item{bigmark}{The separator to the left of the decimal separator}
}
\value{
A list of shorthand positions in a vector
}
\description{
\code{shrthnd_list()} generates a lookup table of shorthand markers in a vector,
either a character vector containing shorthand or a \code{shrthnd_num()} vector.
}
\examples{
x <- c("12", "34.567", "[c]", "NA", "56.78[e]", "78.9", "90.123[e]")
shrthnd_list(x)

sh_x <- shrthnd_num(x)
sh_x
shrthnd_list(sh_x)
}
\seealso{
\code{\link[=is_shrthnd_list]{is_shrthnd_list()}} \code{\link[=shrthnd_num]{shrthnd_num()}}
}
