% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrthand_num.R
\name{shrthnd_num}
\alias{shrthnd_num}
\title{Process a character vector containing shorthand}
\usage{
shrthnd_num(x, shorthand, na_values = "NA", digits = 2L)
}
\arguments{
\item{x}{A character vector of numeric values with shorthand}

\item{shorthand}{A character vector of shorthand values}

\item{na_values}{A character vector of NA values}

\item{digits}{The number of digits for formatting numbers}
}
\value{
A shrthnd_num vector
}
\description{
Process a character vector containing shorthand
}
\examples{
x <- c("12.345", "34.567", "[c]", "NA", "56.78[e]", "78.9", "90.123[e]")
shrthnd_num(x, c("[c]", "[e]"))
}
