% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrthnd_num.R
\name{shrthnd_num}
\alias{shrthnd_num}
\title{Process a character vector containing shorthand}
\usage{
shrthnd_num(
  x,
  shorthand = NULL,
  na_values = "NA",
  digits = 2L,
  paren_nums = c("negative", "strip")
)
}
\arguments{
\item{x}{A character vector of numeric values with shorthand}

\item{shorthand}{A character vector of shorthand values}

\item{na_values}{A character vector of NA values}

\item{digits}{The number of digits for formatting numbers}

\item{paren_nums}{How to handle numbers in parenthesis (e.g. \verb{(12,435.43)}),
defaults to negative as most commonly used in accounting to denote
negative values instead of a minus symbol preceding the value}
}
\value{
A shrthnd_num vector
}
\description{
\code{shrthnd_num()} coerces a character vector containing numeric data values
with non-numeric tags into a numeric vector while also retaining the
tags.
}
\details{
Data stored in documents and publications are regularly annotated with
shorthand and symbols. Often these tags are found in the same container
(e.g. a table or spreadsheet cell) as the value they are associated with,
which requires further cleaning of the vector to extract the numeric values.

A simple approach is to discard the non-numeric components, however these
markers convey information which you may wish to retain. \code{shrthnd_num()}
provides a data type that can store both the numeric data and the marker.

\code{shrthnd_num()} expects character vectors with a numeric component followed
by a non-numeric component, other formats will fail.

\code{shrthnd_num} vectors have one of two subclasses, \code{shrthnd_integer} when
the underlying data is an integer and \code{shrthnd_double} when the underlying
data is a real number.

By default \code{shrthnd_num()} will extract any non-numeric values following
numeric ones and process them as a shorthand tag. However, you can
optionally supply a vector of tags, using the \code{shorthand} argument, if you
wish to validate the extracted tags and only accept vectors with specific
shorthand values.

If the underlying numeric values are real numbers (i.e. a \code{double()} vector)
the \code{digits} argument will be used to format the display of the
\code{shrthnd_dbl} vector (defaults to 2 decimal places).
}
\examples{
x <- c("12", "34.567", "[c]", "NA", "56.78[e]", "78.9", "90.123[e]")
shrthnd_num(x, c("[c]", "[e]"))
}
