% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{shrthnd_tbl}
\alias{shrthnd_tbl}
\title{Add notes to tibbles}
\usage{
shrthnd_tbl(tbl, title = NULL, notes = NULL, source_note = NULL)
}
\arguments{
\item{tbl}{A \code{tibble::tibble()} or object that can be coerced to a tibble.}

\item{title}{A character vector representing the title of the tbl}

\item{notes}{A character vector of notes relating to the title}

\item{source_note}{A character vector for a source note}
}
\value{
A tibble with shrthnd annotations
}
\description{
Add notes to tibbles
}
\examples{
x <- c("12", "34.567", "[c]", "NA", "56.78[e]", "78.9", "90.123[e]")
sh_x <- shrthnd_num(x, c("[c]", "[e]"))
tbl <- tibble::tibble(x = x, sh_x = sh_x)
shrthnd_tbl(
  tbl,
  title = "Example table",
  notes = c("Note 1", "Note 2"),
  source_note = "Shrthnd documentation, 2023"
)
}
\seealso{
\code{\link[=is_shrthnd_tbl]{is_shrthnd_tbl()}}


\code{\link{note_to_title}()},
\code{\link{shrthnd_title}()},
\code{\link{zap_shrthnd}()}
}
\concept{tbl}
