% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_locations.R
\name{tag_match}
\alias{tag_match}
\alias{tag_in}
\alias{where_tag}
\alias{any_tag}
\alias{is_na_tag}
\alias{is_na_both}
\alias{locate_tag}
\alias{locate_any_tag}
\alias{locate_no_tag}
\title{Get tag locations}
\usage{
tag_match(x, tag)

tag_in(x, tag)

where_tag(x, tag)

any_tag(x)

is_na_tag(x)

is_na_both(x)

locate_tag(x, tag)

locate_any_tag(x)

locate_no_tag(x)
}
\arguments{
\item{x}{A \code{shrthnd_num()} vector}

\item{tag}{A single tag to locate}
}
\value{
For \code{tag_match()}, \code{locate_tag()}, \code{locate_any_tag()} and
\code{locate_no_tag()} an integer vector. For \code{tag_in()}, \code{where_tag()},
\code{any_tag()}, \code{is_na_tag()} and \code{is_na_both()} a logical vector.
}
\description{
Base R's matching and location functions will work directly with the
numeric component of a \code{shrthnd_num()} vector, these functions provide
the same functionality but applied to the tag component.
}
\details{
\code{tag_match()} and \code{tag_in()} are wrappers around \code{vctrs::vec_match()} and
\code{vctrs::vec_in()} and thus equivalent to \code{match()} and \code{\%in\%} as applied
to the tag components of a \code{shrthnd_num()}. \code{tag_match()} will return an
integer vector showing the first location of the tag provided, \code{tag_in()}
will return \code{TRUE} or \code{FALSE} depending on whether the tag is in the
vector's shorthand.

\code{where_tag()} is equivalent to computing \code{tags == tag}, \code{any_tag()} is
equivalent to \code{!is.na(tags)}. Using \code{is.na()} on a \code{shrthnd_num()} will
assess if the numeric component is missing, \code{is_na_tag()} is equivalent to
\code{is.na(tags)}, \code{is_na_both()} tests if both the numeric and tag components
of a \code{shrthnd_num()} are missing. They return a logical
vector the same length as \code{x}.

\code{locate_tag()}, \code{locate_any_tag()}, \code{located_no_tag()} are equivalent to
passing the return values of \code{where_tag()}, \code{any_tag()} and \code{is_na_tag()} to
\code{which()}. They return an integer vector the same length as \code{x}.
}
\examples{
x <- c("12", "34.567", "[c]", "NA", "56.78[e]", "78.9", "90.123[e]")
sh_x <- shrthnd_num(x, c("[c]", "[e]"))
shrthnd_tags(sh_x)

tag_match(sh_x, "[e]")

tag_in(sh_x, "[e]")

where_tag(sh_x, "[e]")

any_tag(sh_x)

is_na_tag(sh_x)

is_na_both(sh_x)

locate_tag(sh_x, "[e]")

locate_any_tag(sh_x)

locate_no_tag(sh_x)
}
\seealso{

\code{\link{as_shrthnd}()},
\code{\link{is_shrthnd_num}()},
\code{\link{make_shrthnd_num}()},
\code{\link{shrthnd_coercion}},
\code{\link{shrthnd_maths}},
\code{\link{shrthnd_num}()},
\code{\link{shrthnd_tags}()}
}
\concept{num}
